+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/conf/WaggleDanceConfiguration.java
@@ -45,6 +45,8 @@ public class WaggleDanceConfiguration {
   private int thriftServerStopTimeoutValInSeconds = 60;
   private int thriftServerRequestTimeout = 20;
   private TimeUnit thriftServerRequestTimeoutUnit = TimeUnit.SECONDS;
+  private int statusPollingDelay = 5;
+  private TimeUnit statusPollingDelayTimeUnit = TimeUnit.MINUTES;
 
   public Integer getPort() {
     return port;
@@ -116,6 +118,22 @@ public class WaggleDanceConfiguration {
 
   public void setThriftServerRequestTimeoutUnit(TimeUnit thriftServerRequestTimeoutUnit) {
     this.thriftServerRequestTimeoutUnit = thriftServerRequestTimeoutUnit;
+  }
+
+  public int getStatusPollingDelay() {
+    return statusPollingDelay;
+  }
+
+  public void setStatusPollingDelay(int statusPollingDelay) {
+    this.statusPollingDelay = statusPollingDelay;
+  }
+
+  public TimeUnit getStatusPollingDelayTimeUnit() {
+    return statusPollingDelayTimeUnit;
+  }
+
+  public void setStatusPollingDelayTimeUnit(TimeUnit statusPollingDelayTimeUnit) {
+    this.statusPollingDelayTimeUnit = statusPollingDelayTimeUnit;
   }
 
 }