+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/context/CommonBeans.java
@@ -25,10 +25,12 @@ import com.hotels.bdp.waggledance.client.CloseableThriftHiveMetastoreIfaceClient
 import com.hotels.bdp.waggledance.client.DefaultMetaStoreClientFactory;
 import com.hotels.bdp.waggledance.client.tunnelling.TunnelingMetaStoreClientFactory;
 import com.hotels.bdp.waggledance.conf.WaggleDanceConfiguration;
+import com.hotels.bdp.waggledance.core.federation.service.PopulateStatusFederationService;
 import com.hotels.bdp.waggledance.mapping.model.ASTQueryMapping;
 import com.hotels.bdp.waggledance.mapping.model.QueryMapping;
 import com.hotels.bdp.waggledance.mapping.service.PrefixNamingStrategy;
 import com.hotels.bdp.waggledance.mapping.service.impl.LowerCasePrefixNamingStrategy;
+import com.hotels.bdp.waggledance.mapping.service.impl.PollingFederationService;
 
 @org.springframework.context.annotation.Configuration
 public class CommonBeans {
@@ -61,6 +63,12 @@ public class CommonBeans {
   @Bean
   public QueryMapping queryMapping() {
     return ASTQueryMapping.INSTANCE;
+  }
+
+  @Bean
+  public PollingFederationService pollingFederationService(
+      PopulateStatusFederationService populateStatusFederationService) {
+    return new PollingFederationService(populateStatusFederationService);
   }
 
 }